Refactor wallet folder and move wallet history store to component folder
