+++ src/scenes/Wallet/ActionButtonHeader/EncryptDialog/index.js
