+++ src/scenes/Wallet/Balances/DepositDialog/styles.js
